Trigger check implemented

+++ src/triggers/RelationshipContacts.trigger
@@ -31,7 +31,7 @@ trigger RelationshipContacts on Contact (after delete) {
 
 	public enum triggerAction {beforeInsert, beforeUpdate, beforeDelete, afterInsert, afterUpdate, afterDelete, afterUndelete}
 
-    if( Trigger.isAfter && Trigger.isDelete ){
+    if(Trigger.isAfter && Trigger.isDelete && !Relationships_Utils.getRelationshipSettings().DISABLE_RelationshipContacts_trigger__c){
         Relationships.deleteEmptyRelationships();
     }
 }
